\name{rvle}
\alias{rvle.addIntegerCondition}
\title{Add an integer to the specified condition port.}
\description{Add an integer to the specified condition port.}
\usage{rvle.addIntegerCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an integer}
}
\keyword{models}
\keyword{optimize}

