\name{rvle}
\alias{rvle.addRealCondition}
\title{Add a real to the specified condition port.}
\description{Add a real to the specified condition port.}
\usage{rvle.addRealCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a real}
}
\keyword{models}
\keyword{optimize}

