\name{rvle}
\alias{rvle.addStringCondition}
\title{Add a string to the specified condition port.}
\description{Add an string to the specified condition port.}
\usage{rvle.addStringCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a string}
}
\keyword{models}
\keyword{optimize}

