\name{rvle}
\alias{rvle.clearConditionPort}
\title{Clear all values from the specified condition port.}
\description{Clear all values from the specified condition port.}
\usage{rvle.clearConditionPort(self, condition, port)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
}
\keyword{models}
\keyword{optimize}

