\name{rvle}
\alias{rvle.condition_add_integer}
\title{Add an integer to the specified condition port.}
\description{Add an integer to the specified condition port.}
\usage{condition_add_integer(f, condition, port, value)}
\arguments{
        \item{f}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an integer}
}
\keyword{models}
\keyword{optimize}

