\name{rvle}
\alias{rvle.condition_add_real}
\title{Add a real to the specified condition port.}
\description{Add a real to the specified condition port.}
\usage{condition_add_real(f, condition, port, value)}
\arguments{
        \item{f}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a real}
}
\keyword{models}
\keyword{optimize}

