\name{rvle}
\alias{rvle.condition_add_string}
\title{Add a string to the specified condition port.}
\description{Add an string to the specified condition port.}
\usage{condition_add_string(f, condition, port, value)}
\arguments{
        \item{f}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a string}
}
\keyword{models}
\keyword{optimize}

