\name{rvle}
\alias{rvle.condition_port_list}
\title{Get the ports attached to a specific condition.}
\description{Get the ports attached to a specific condition.}
\usage{condition_port_list(f, condition)}
\arguments{
	\item{f}{handle of the rvle}
	\item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}

