\name{rvle}
\alias{rvle.condition_port_list_size}
\title{Get the number of port for a specific condition.}
\description{Get the number of port for a specific condition.}
\usage{condition_port_list_size(f, condition)}
\arguments{
	\item{f}{handle of the rvle}
	\item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}
