\name{rvle}
\alias{rvle.condition_set_integer}
\title{Set a integer to the specified condition port.}
\description{Set a integer to the specified condition port.}
\usage{condition_set_integer(x, condition, port, value)}
\arguments{
        \item{x}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an integer}
}
\keyword{models}
\keyword{optimize}

