\name{rvle}
\alias{rvle.condition_set_real}
\title{Set a real to the specified condition port.}
\description{Set a real to the specified condition port.}
\usage{condition_set_real(x, condition, port, value)}
\arguments{
        \item{x}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a real}
}
\keyword{models}
\keyword{optimize}

