\name{rvle}
\alias{rvle.condition_set_string}
\title{Set a string to the specified condition port.}
\description{Set a string to the specified condition port.}
\usage{condition_set_string(x, condition, port, value)}
\arguments{
        \item{x}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an string}
}
\keyword{models}
\keyword{optimize}

