\name{rvle}
\alias{rvle.condition_show}
\title{Show the contains of a specified condition name, port name couple}
\description{Show the contains of a specified condition}
\usage{condition_show(f, condition, port)}
\arguments{
        \item{f}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
}
\keyword{models}
\keyword{optimize}

