\name{rvle}
\alias{rvle.experiment_set_duration}
\title{Set the duration of the experiment.}
\description{Set the duration of the experiment.}
\usage{experiment_set_duration(x, value)}
\arguments{
        \item{x}{handle of the rvle}
        \item{value}{a real to set the duration}
}
\keyword{models}
\keyword{optimize}
