\name{rvle}
\alias{rvle.experiment_set_seed}
\title{Set the seed of the experiment.}
\description{Set the seed of the experiment.}
\usage{experiment_set_seed(x, value)}
\arguments{
        \item{x}{handle of the rvle}
        \item{value}{an integer to set the seed}
}
\keyword{models}
\keyword{optimize}
