\name{rvle}
\alias{rvle.experiment_total_combination}
\title{Assign a total experimental frame with specific seed to build seeds of
    simulations and the number of repliquas.}
\description{Assign a total experimental frame with specific seed to build
    seeds of simulations and the number of repliquas.}
\usage{experiment_total_combination(x, seed, repliquas)}
\arguments{
        \item{x}{handle of the rvle}
        \item{seed}{an integer to set the seed of random seed}
        \item{repliquas}{an integer to set the number of repliquas}
}
\keyword{models}
\keyword{optimize}
