\name{rvle}
\alias{rvle.getConditionPortsSize}
\title{Get the number of port for a specific condition.}
\description{Get the number of port for a specific condition.}
\usage{rvle.getConditionPortsSize(self, condition)}
\arguments{
	\item{self}{handle of the rvle}
	\item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}
