\name{rvle}
\alias{rvle.getConditionPortValues}
\title{Show the contains of a specified condition name, port name couple}
\description{Show the contains of a specified condition}
\usage{rvle.getConditionPortValues(self, condition, port)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
}
\keyword{models}
\keyword{optimize}

