\name{rvle}
\alias{rvle.getOutputPlugin}
\title{Get the plugin output of a view.}
\description{Get the plugin output of a view.}
\usage{rvle.getOutputPlugin(self, viewname)}
\arguments{
        \item{self}{handle of the rvle}
        \item{viewname}{the name of the view.}
}
\keyword{models}
\keyword{optimize}
