\name{rvle}
\alias{rvle.listConditionPorts}
\title{Get the ports attached to a specific condition.}
\description{Get the ports attached to a specific condition.}
\usage{rvle.listConditionPorts(self, condition)}
\arguments{
	\item{self}{handle of the rvle}
	\item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}

