\name{rvle}
\alias{rvle.open}
\title{Load a new Vpz file.}
\description{Load a new Vpz file. The Vpz file can either be global, or
  relative to a package. In the first case, only the first argument is
  to be set to a file path. In the second case the first argument need
  to be set to a file name, and the second one to the package name of
  the experiment.}
\usage{open(file,[pkg])}
\arguments{
  \item{file}{character string giving the name of a file to load or
    create.}
  \item{pkg}{character string giving the name of the package.}
}
\keyword{models}
\keyword{optimize}

