\name{rvle}
\alias{rvle.runManager}
\title{Use an experimental frame and launch simulation.}
\description{Use an experimental frame from the rvle handle and launch
simulation for each instance of the experimental frame. Returns a matrix
of list of reals matrix.}
\usage{rvle.runManager(self)}
\arguments{
	\item{self}{handle of rvle}
}
\keyword{models}
\keyword{optimize}

