\name{rvle}
\alias{rvle.run_manager_cluster}
\title{Use an experimental frame and launch simulation onto grid computing.}
\description{Use an experimental frame from the rvle handle and launch
simulation for each instance of the experimental frame onto a grid computing.
Returns a matrix of list of dataframes.}
\usage{run_manager_cluster(f)}
\arguments{
	\item{f}{handle of rvle}
}
\keyword{models}
\keyword{optimize}


