\name{rvle}
\alias{rvle.run_manager_thread}
\title{Use an experimental frame and launch simulation onto local thread.}
\description{Use an experimental frame from the rvle handle and launch
simulation for each instance of the experimental frame onto local threads.
Returns a matrix of list of dataframes.}
\usage{run_manager_thread(f, th)}
\arguments{
	\item{f}{handle of rvle}
	\item{th}{number of threads, one tread by instance of the experimental
frames}
}
\keyword{models}
\keyword{optimize}



