\name{rvle}
\alias{rvle.runManagerThreadMatrix}
\title{Use an experimental frame and launch simulation onto local thread.}
\description{Use an experimental frame from the rvle handle and launch
simulation for each instance of the experimental frame onto local threads.
Returns a matrix of list of matrix of reals.}
\usage{rvle.runManagerThreadMatrix(self, th)}
\arguments{
	\item{self}{handle of rvle}
	\item{th}{number of threads, one tread by instance of the experimental
frames}
}
\keyword{models}
\keyword{optimize}




