\name{rvle}
\alias{rvle.save}
\title{Write a vpz file under the specified filename.}
\description{Write a vpz file under the specified filename.}
\usage{rvle.save(self, filename)}
\arguments{
    \item{self}{handle of the rvle}
    \item{filename}{string representation of the filename where store vpz file}
}
\keyword{models}
\keyword{optimize}

