\name{rvle}
\alias{rvle.setDuration}
\title{Set the duration of the experiment.}
\description{Set the duration of the experiment.}
\usage{rvle.setDuration(self, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{value}{a real to set the duration}
}
\keyword{models}
\keyword{optimize}
