\name{rvle}
\alias{rvle.setIntegerCondition}
\title{Set a integer to the specified condition port.}
\description{Set a integer to the specified condition port.}
\usage{rvle.setIntegerCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an integer}
}
\keyword{models}
\keyword{optimize}

