\name{rvle}
\alias{rvle.setOutputPlugin}
\title{Set the plugin output of a view.}
\description{Set the plugin output of a view.}
\usage{rvle.setOutputPlugin(self, viewname, pluginname)}
\arguments{
        \item{self}{handle of the rvle}
        \item{viewname}{the name of the view.}
        \item{pluginname}{the name of the plugin}
}
\keyword{models}
\keyword{optimize}
