\name{rvle}
\alias{rvle.setStringCondition}
\title{Set a string to the specified condition port.}
\description{Set a string to the specified condition port.}
\usage{rvle.setStringCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an string}
}
\keyword{models}
\keyword{optimize}

