\name{rvle}
\alias{rvle.setTotalCombination}
\title{Assign a total experimental frame with specific seed to build seeds of
    simulations and the number of repliquas.}
\description{Assign a total experimental frame with specific seed to build
    seeds of simulations and the number of repliquas.}
\usage{rvle.setTotalCombination(self, seed, repliquas)}
\arguments{
        \item{self}{handle of the rvle}
        \item{seed}{an integer to set the seed of random seed}
        \item{repliquas}{an integer to set the number of repliquas}
}
\keyword{models}
\keyword{optimize}
