\name{rvle}
\alias{rvle.run_manager_matrix}
\title{Use an experimental frame and launch simulation.}
\description{Use an experimental frame from the rvle handle and launch
simulation for each instance of the experimental frame. Returns a matrix of list
of dataframes.}
\usage{run_manager(f)}
\arguments{
	\item{f}{handle of rvle}
}
\keyword{models}
\keyword{optimize}


